remove legacy trainers
